+++ src/runtime/dictionary/icons.ts
@@ -7,7 +7,7 @@
 import ArrowToTheLeftIcon from '@bitrix24/b24icons-vue/actions/ArrowToTheLeftIcon'
 import ArrowToTheRightIcon from '@bitrix24/b24icons-vue/actions/ArrowToTheRightIcon'
 import CheckIcon from '@bitrix24/b24icons-vue/outline/CheckLIcon'
-import ChevronUpIcon from '@bitrix24/b24icons-vue/actions/ChevronUpIcon'
+import ChevronTopLIcon from '@bitrix24/b24icons-vue/outline/ChevronTopLIcon'
 import ChevronToTheLeftIcon from '@bitrix24/b24icons-vue/outline/ChevronLeftLIcon'
 import ChevronToTheRightIcon from '@bitrix24/b24icons-vue/outline/ChevronRightLIcon'
 import DoubleShevronsRightIcon from '@bitrix24/b24icons-vue/actions/DoubleShevronsRightIcon'
@@ -31,7 +31,7 @@ export default {
   chevronDown: ChevronDownLIcon,
   chevronLeft: ChevronToTheLeftIcon,
   chevronRight: ChevronToTheRightIcon,
-  chevronUp: ChevronUpIcon,
+  chevronUp: ChevronTopLIcon,
   close: Cross30Icon,
   ellipsis: DotsIcon,
   external: OpenIn50Icon,