+++ CHANGELOG.md
@@ -185,6 +185,13 @@
   - @deprecate color
     - default, danger, success, warning, primary, secondary, collab, ai
   - use `Badge` as tag
+* **InputNumber**
+  - fix color, size
+  - add color
+    - air-primary, air-primary-success, air-primary-alert, air-primary-copilot, air-primary-warning
+  - @deprecate color
+    - default, danger, success, warning, primary, secondary, collab, ai
+  - use `Badge` as tag
 ---
 We change btn colors: